# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecom Libya Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Product {
  id: string;
  /** URL-friendly identifier generated from product name */
  slug?: string | null;
  name: string;
  nameAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  compareAtPrice?: number | null;
  imageUrl?: string | null;
  images: string[];
  categoryId?: string | null;
  categoryName?: string | null;
  stock: number;
  sku?: string | null;
  featured: boolean;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateProductInput {
  name: string;
  nameAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  compareAtPrice?: number | null;
  imageUrl?: string | null;
  images?: string[];
  categoryId?: string | null;
  stock: number;
  sku?: string | null;
  featured?: boolean;
  active?: boolean;
}

export interface UpdateProductInput {
  name?: string;
  nameAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  compareAtPrice?: number | null;
  imageUrl?: string | null;
  images?: string[];
  categoryId?: string | null;
  stock?: number;
  sku?: string | null;
  featured?: boolean;
  active?: boolean;
}

export interface Category {
  id: string;
  name: string;
  nameAr: string;
  description?: string | null;
  imageUrl?: string | null;
  productCount: number;
  createdAt: string;
}

export interface CreateCategoryInput {
  name: string;
  nameAr: string;
  description?: string | null;
  imageUrl?: string | null;
}

export interface CartItem {
  id: string;
  productId: string;
  productName: string;
  productNameAr: string;
  productImage?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartInput {
  sessionId: string;
  productId: string;
  quantity: number;
}

export interface UpdateCartItemInput {
  quantity: number;
}

export interface OrderItem {
  id: string;
  productId: string;
  productName: string;
  productNameAr: string;
  productImage?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: string;
  orderNumber: string;
  status: OrderStatus;
  customerName: string;
  customerPhone: string;
  customerCity: string;
  customerAddress: string;
  items: OrderItem[];
  subtotal: number;
  shippingFee: number;
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateOrderInput {
  sessionId: string;
  customerName: string;
  customerPhone: string;
  customerCity: string;
  customerAddress: string;
  notes?: string | null;
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export type AdminStatsTopProductsItem = {
  productId: string;
  productName: string;
  totalSold: number;
  revenue: number;
};

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  totalProducts: number;
  lowStockProducts: number;
  recentOrders: Order[];
  topProducts: AdminStatsTopProductsItem[];
}

export type GetProductsParams = {
  categoryId?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  inStock?: boolean;
  page?: number;
  limit?: number;
};

export type GetCartParams = {
  sessionId: string;
};

export type GetOrdersParams = {
  status?: string;
  page?: number;
  limit?: number;
};