# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecom Libya Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: string;
  productId: string;
  productName: string;
  productNameAr: string;
  productImage?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}