# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ecom Libya Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  /** URL-friendly identifier generated from product name */
  slug?: string | null;
  name: string;
  nameAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  compareAtPrice?: number | null;
  imageUrl?: string | null;
  images: string[];
  categoryId?: string | null;
  categoryName?: string | null;
  stock: number;
  sku?: string | null;
  featured: boolean;
  active: boolean;
  createdAt: Date;
  updatedAt: Date;
}